+++ services/device_gateway/include/device_gateway/gateway_config.h
@@ -9,6 +9,7 @@ namespace deviceops::gateway {
 
 struct GatewayConfig {
     std::string gateway_id = "device-gateway-001";
+    int rpc_port = 9101;
     tewmqtt::MqttConfig mqtt;
     std::vector<std::string> subscribe_topics;
 };